+++ tests/examples/path_filter/plan.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 from stepup.core.api import amend
-from stepup.core.utils import filter_dependencies
+from stepup.core.extapi import filter_dependencies
 
 paths = filter_dependencies(["bar.py", "venv/foo.py"])
 amend(vol="paths.txt")